requirements.txt
selenium==4.18.1
python-telegram-bot==21.0.1
Pillow==10.3.0
